requirements.txt
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
Flask==3.1.3
Flask-Dance==7.1.0
Flask-Limiter==4.1.1
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
oauthlib==3.3.1
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
pyClamd==0.4.0
pycparser==3.0
pycryptodome==3.23.0
Pygments==2.19.2
pyotp==2.9.0
pytest==9.0.2
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
URLObject==3.0.0
waitress==3.0.2
Werkzeug==3.1.6
wrapt==2.1.1
WTForms==3.2.1
